Backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
mysql-connector-python==8.2.0
pymongo==4.6.0
python-dotenv==1.0.0
Werkzeug==3.0.1
requests==2.32.5
pytest==7.4.3
pytest-flask==1.3.0
